Changes addressing the reviewer’s comments

+++ sdp/processors/manage_files/remove.py
@@ -14,6 +14,12 @@
 
 import os
 import shutil
+from pathlib import Path
+from collections import Counter
+from tqdm import tqdm
+import itertools
+from tqdm.contrib.concurrent import process_map
+
 
 from sdp.logging import logger
 from sdp.processors.base_processor import DataEntry, BaseParallelProcessor
@@ -30,6 +36,8 @@ class RemoveFiles(BaseParallelProcessor):
         filepath_field (str): The key in the data entry that holds the path to the file or directory to remove.
         
         drop_filepath_field (bool): Whether to remove the filepath field from the resulting data entry. Defaults to True.
+
+        recursive (bool): Whether to recursively remove files from directories. Defaults to False.
         
         **kwargs: Additional arguments passed to the BaseParallelProcessor.
     
@@ -54,11 +62,49 @@ class RemoveFiles(BaseParallelProcessor):
     def __init__(self,
                 filepath_field: str,
                 drop_filepath_field: bool = True,
+                recursive: bool = False,
                 **kwargs):
 
         super().__init__(**kwargs)
         self.filepath_field = filepath_field
         self.drop_filepath_field = drop_filepath_field
+        self.recursive = recursive
+    
+    def _count_files(self, data_entry):
+        """
+        Count the number of files to be removed.
+        """
+        filepath = Path(data_entry[self.filepath_field])
+        if filepath.is_dir():
+            if self.recursive:
+                file_counter = Counter(f.suffix for f in filepath.iterdir() if f.is_file())
+            else:
+                raise IsADirectoryError(f"Directory {filepath} is not empty and recursive is False")
+        else:
+            file_counter = Counter({filepath.suffix : 1})
+        return file_counter
+
+    def prepare(self):
+        """
+        Prepare the processor by counting the number of files to be removed.
+        """
+        file_counter = Counter()
+        for manifest_chunk in self._chunk_manifest():
+            chunk_counts = itertools.chain(
+                process_map(
+                    self._count_files,
+                    manifest_chunk,
+                    max_workers=self.max_workers,
+                    chunksize=self.chunksize,
+                    desc="Counting files to remove",
+                )
+            )
+            for entry_file_counter in chunk_counts:
+                file_counter += entry_file_counter
+
+        print(f"Total files to remove: {sum(file_counter.values())}")
+        for extension, count in file_counter.items():
+            logger.info(f"{extension}\t\t{count}")
 
     def process_dataset_entry(self, data_entry):
         """
@@ -74,7 +120,10 @@ class RemoveFiles(BaseParallelProcessor):
 
         # Remove the target path from the filesystem
         if os.path.isdir(filepath):
-            shutil.rmtree(filepath)  # Recursively delete directory
+            if self.recursive:
+                shutil.rmtree(filepath)  # Recursively delete directory
+            else:
+                raise IsADirectoryError(f"Directory {filepath} is not empty and recursive is False")
         else:
             os.remove(filepath)  # Delete a single file
